# docs/protocol/local-admin-api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Private Sync Gate Local Administration API",
    "version": "1.0.0",
    "description": "Loopback-only, owner-controlled administration API for the local daemon."
  },
  "servers": [
    { "url": "http://127.0.0.1:47820", "description": "Local daemon loopback listener" }
  ],
  "paths": {
...
    "/api/v1/devices": {
      "get": {
        "operationId": "listDevices",
        "summary": "List paired devices",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "description": "Device page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DevicePage" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
...
    "/api/v1/pairing/invitations": {
      "post": {
        "operationId": "createPairingInvitation",
        "summary": "Create a short-lived pairing invitation",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvitationRequest" } } } },
        "responses": { "201": { "description": "Invitation created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invitation" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/pairing/inspect": {
      "post": {
        "operationId": "inspectPairingInvitation",
        "summary": "Inspect an invitation without accepting it",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EncodedInvitation" } } } },
        "responses": { "200": { "description": "Invitation metadata", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvitationInspection" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/pairing/acceptances": {
      "post": {
        "operationId": "acceptPairingInvitation",
        "summary": "Accept a pairing invitation",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AcceptanceRequest" } } } },
        "responses": { "200": { "description": "Pairing accepted or already accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Acceptance" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "409": { "$ref": "#/components/responses/Conflict" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "local-admin-token" },
      "browserSession": { "type": "apiKey", "in": "cookie", "name": "syncgate_browser_session" }
    },
    "parameters": {
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 } },
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string", "maxLength": 512 } },
...
    },
    "responses": {
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Missing or invalid local credential", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Conflict": { "description": "Command conflicts with current state", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unavailable": { "description": "Daemon service unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message"], "properties": { "code": { "type": "string" }, "message": { "type": "string", "maxLength": 512 }, "request_id": { "type": "string", "maxLength": 128 } } } } },
...
      "PageInfo": { "type": "object", "required": ["limit", "has_more"], "properties": { "limit": { "type": "integer", "minimum": 1, "maximum": 200 }, "next_cursor": { "type": "string", "maxLength": 512 }, "has_more": { "type": "boolean" } } },
...
      "Device": { "type": "object", "required": ["id", "name", "fingerprint", "state"], "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "fingerprint": { "type": "string" }, "state": { "type": "string", "enum": ["paired", "revoked", "offline"] }, "last_seen_at": { "type": "string", "format": "date-time" } } },
...
      "DevicePage": { "type": "object", "required": ["items", "page"], "properties": { "items": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/Device" } }, "page": { "$ref": "#/components/schemas/PageInfo" } } },
...
      "InvitationRequest": { "type": "object", "required": ["ttl_seconds"], "additionalProperties": false, "properties": { "ttl_seconds": { "type": "integer", "minimum": 1, "maximum": 86400 }, "capabilities": { "type": "array", "maxItems": 16, "items": { "type": "string", "maxLength": 64 } } } },
      "Invitation": { "type": "object", "required": ["invitation", "fingerprint", "one_time_code", "expires_at"], "additionalProperties": false, "properties": { "invitation": { "type": "string", "maxLength": 65536 }, "fingerprint": { "type": "string" }, "one_time_code": { "type": "string" }, "expires_at": { "type": "string", "format": "date-time" } } },
      "EncodedInvitation": { "type": "object", "required": ["invitation"], "additionalProperties": false, "properties": { "invitation": { "type": "string", "minLength": 1, "maxLength": 65536 } } },
      "InvitationInspection": { "type": "object", "required": ["valid", "invite_id", "device_id", "device_name", "fingerprint", "created_at", "expires_at", "capabilities"], "additionalProperties": false, "properties": { "valid": { "type": "boolean", "const": true }, "invite_id": { "type": "string" }, "device_id": { "type": "string" }, "device_name": { "type": "string", "maxLength": 256 }, "fingerprint": { "type": "string" }, "created_at": { "type": "string", "format": "date-time" }, "expires_at": { "type": "string", "format": "date-time" }, "capabilities": { "type": "array", "maxItems": 16, "items": { "type": "string" } } } },
      "PairingGrant": { "type": "object", "required": ["share_id", "capabilities"], "additionalProperties": false, "properties": { "share_id": { "type": "string", "minLength": 1, "maxLength": 256 }, "capabilities": { "type": "array", "minItems": 1, "maxItems": 16, "uniqueItems": true, "items": { "type": "string" } }, "lan_only": { "type": "boolean", "default": true } } },
      "ControlPlaneGrant": { "type": "object", "required": ["read_status", "ttl_seconds"], "additionalProperties": false, "properties": { "read_status": { "type": "boolean", "const": true }, "ttl_seconds": { "type": "integer", "minimum": 60, "maximum": 2592000 } } },
      "AcceptanceRequest": { "type": "object", "required": ["invitation", "expected_fingerprint", "one_time_code", "grants"], "additionalProperties": false, "properties": { "invitation": { "type": "string", "minLength": 1, "maxLength": 65536 }, "expected_fingerprint": { "type": "string", "minLength": 1, "maxLength": 256 }, "one_time_code": { "type": "string", "minLength": 1, "maxLength": 128 }, "grants": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/PairingGrant" } }, "control_plane_grant": { "$ref": "#/components/schemas/ControlPlaneGrant" } } },
      "Acceptance": { "type": "object", "required": ["accepted", "already_accepted", "device_id", "device_name", "fingerprint"], "additionalProperties": false, "properties": { "accepted": { "type": "boolean", "const": true }, "already_accepted": { "type": "boolean" }, "device_id": { "type": "string" }, "device_name": { "type": "string" }, "fingerprint": { "type": "string" } } },
...
    }
  }
}