# internal/storage/models.go
package storage

import (
	"time"

	"syncgate/internal/core"
)

type TrustState string

const (
	TrustPending TrustState = "pending"
	TrustTrusted TrustState = "trusted"
	TrustRevoked TrustState = "revoked"
)

type Device struct {
	ID          core.DeviceID
	DisplayName string
	PublicKey   []byte
	Fingerprint string
	TrustState  TrustState
	CreatedAt   time.Time
	UpdatedAt   time.Time
	LastSeenAt  time.Time
}

type ShareMode string

const (
	ShareSendOnce      ShareMode = "send_once"
	ShareOneWaySource  ShareMode = "one_way_source"
	ShareOneWayTarget  ShareMode = "one_way_target"
	ShareUploadOnly    ShareMode = "upload_only"
	ShareReadOnly      ShareMode = "read_only"
	ShareTwoWayPlanned ShareMode = "two_way_planned"
)

type Share struct {
	ID                   core.ShareID
	Name                 string
	RootPath             string
	Mode                 ShareMode
	CasePolicy           string
	VersionPolicy        string
	DeletionLimitCount   int
	DeletionLimitPercent int
	CreatedAt            time.Time
	UpdatedAt            time.Time
}

// Tombstone records an accepted source deletion. Tombstones are immutable
// history; active propagation is determined by the current file-index pointer.
type Tombstone struct {
	ID                  core.TombstoneID
	ShareID             core.ShareID
	RelativePath        string
	DeletedByDeviceID   core.DeviceID
	BaseRevisionID      core.RevisionID
	TombstoneRevisionID core.RevisionID
	DeletedAt           time.Time
	ExpiresAt           time.Time
}

type AuditEvent struct {
	ID            string
	EventName     string
	DeviceID      core.DeviceID
	PeerDeviceID  core.DeviceID
	ShareID       core.ShareID
	TransferID    core.TransferID
	RevisionID    core.RevisionID
	TransportType string
	Severity      string
	Metadata      map[string]string
	OccurredAt    time.Time
}

type PairingAcceptance struct {
	InviteID          string
	Device            Device
	Permissions       []core.SharePermission
	ControlPlaneGrant *ControlPlaneGrant
	AuditEvent        AuditEvent
	AcceptedAt        time.Time
}

type PairingAcceptanceResult struct {
	AlreadyAccepted bool
}

type PairingRevocation struct {
	DeviceID   core.DeviceID
	AuditEvent AuditEvent
	RevokedAt  time.Time
}

type PairingRevocationResult struct {
	AlreadyRevoked bool
}

type AuthenticatedOneWayWork struct {
	Transfer             core.Transfer
	Job                  core.OneWayJob
	RequiredCapabilities []core.Capability
	Remote               bool
}

type AuthenticatedOneWayWorkResult struct {
	AlreadyCreated bool
}